Align velocity tracker
