+++ compose/ui/ui/src/notMobileMain/kotlin/androidx/compose/ui/input/pointer/util/PlatformVelocityTracker.notMobile.kt
@@ -14,7 +14,6 @@
  * limitations under the License.
  */
 
-package androidx.compose.foundation.gestures
+package androidx.compose.ui.input.pointer.util
 
-internal actual fun createDraggableVelocityTracker(): DraggableVelocityTracker =
-    DefaultDraggableVelocityTracker()
+internal actual fun PlatformVelocityTracker(): PlatformVelocityTracker = DefaultVelocityTracker()
